+++ RocketTVCSim/src/Control/Thrust/PID/PIDNew.h
@@ -14,25 +14,25 @@ public:
     /// @param  fc
     ///         Cutoff frequency of derivative EMA filter (Hertz),
     ///         zero to disable the filter entirely
-    PIDNew(float kp, float ki, float kd, float sample_time, float maxOut );
+    PIDNew(double kp, double ki, double kd, double sample_time, double maxOut );
 
     /// Update the controller: given the current position, compute the control
     /// action.
-    float update(float input, float currentTime);
+    double update(double input, double currentTime);
     /// Set the reference/target/setpoint of the controller.
-    void setSetpoint(float setpoint);
+    void setSetpoint(double setpoint);
 private:
-    float sample_time = 1;               ///< Sampling time (seconds)
-    float maxOut = 255;      ///< Maximum control output magnitude
-    float kp = 1;               ///< Proportional gain
-    float ki = 0;            ///< Integral gain times Ts
-    float kd = 0;            ///< Derivative gain divided by Ts
-    float proportional = 0;     ///< Proportional term
-float integral = 0;         ///< Integral term
-float derivative = 0;       ///< Derivative term
-float lastTime = 0;         ///< Time of previous update
-float lastOutput = 0;       ///< Control output of previous update
-float lastInput = 0;        ///< Input of previous update
-    float setpoint = 0;      ///< Position reference.
+    double sample_time = 1;               ///< Sampling time (seconds)
+    double maxOut = 255;      ///< Maximum control output magnitude
+    double kp = 1;               ///< Proportional gain
+    double ki = 0;            ///< Integral gain times Ts
+    double kd = 0;            ///< Derivative gain divided by Ts
+    double proportional = 0;     ///< Proportional term
+double integral = 0;         ///< Integral term
+double derivative = 0;       ///< Derivative term
+double lastTime = 0;         ///< Time of previous update
+double lastOutput = 0;       ///< Control output of previous update
+double lastInput = 0;        ///< Input of previous update
+    double setpoint = 0;      ///< Position reference.
 };
